requirements/requirements.txt
django==3.2
django-settings-export==1.2.1
django-widget-tweaks==1.4.5
django-datatables-view==1.19.1
python-dateutil
django-multiselectfield
psycopg2-binary
django-ckeditor
xlrd
django-model-utils
django-braces
django-resized
django-cleanup
jsonfield
pandas
Werkzeug==2.0.1
pretty-html-table
tqdm
